requirements.txt
python-binance
pandas
numpy
flask
requests
scikit-learn
schedule
python-dotenv